chore: 엔티티 변경 후 Integer 타입으로 수정

+++ src/main/java/efub/eday/edayback/domain/day/title/repository/MemberTitleRepository.java
@@ -9,6 +9,6 @@ import efub.eday.edayback.domain.member.entity.Member;
 import efub.eday.edayback.domain.member.entity.MemberTitle;
 
 @Repository
-public interface MemberTitleRepository extends JpaRepository<MemberTitle, Long> {
+public interface MemberTitleRepository extends JpaRepository<MemberTitle, Integer> {
 	List<MemberTitle> findByMember(Member member);
 }
